// src/common/enums/charge-type.enum.ts
export enum ChargeType {
  MONTHLY_RENT = 'monthly_rent',
  ELECTRICITY_BILL = 'electricity_bill',
  WATER_BILL = 'water_bill',
  MAINTENANCE_CHARGE = 'maintenance_charge',
  OTHER_CHARGE = 'other_charge',
  ADVANCE_PAYMENT = 'advance_payment',
  DISCOUNT_ADJUSTMENT = 'discount_adjustment',
  ADVANCE_REFUND = 'advance_refund',
  RENT_CHANGE = 'rent_change',
}